+++ src/Service/Yaml/Linter/KeysDuplicatedLinter.php
@@ -31,7 +31,7 @@ final class KeysDuplicatedLinter implements LinterInterface
 
     public function getPresets(): array
     {
-        return [LinterRegistry::PRESET_BASE];
+        return [LinterInterface::PRESET_BASE];
     }
 
     public function lint(LintYamlFilterDto $filterDto): ReportBag