+++ src/Service/Yaml/Linter/KeysMissedLinter.php
@@ -29,7 +29,7 @@ final class KeysMissedLinter implements LinterInterface
 
     public function getPresets(): array
     {
-        return [LinterRegistry::PRESET_BASE];
+        return [LinterInterface::PRESET_BASE];
     }
 
     public function lint(LintYamlFilterDto $filterDto): ReportBag